refactor(courtboard): extract CourtCard component
- Create src/courtboard/utils/courtUtils.js with helper functions
  (classForStatus, namesFor, formatTime, computeClock)
- Extract CourtCard to components/CourtCard.jsx
- Remove inline definitions from main.jsx
