+++ src/courtboard/main.jsx
@@ -22,6 +22,10 @@ import './browser-bridge.js';
 // Extracted components
 import { ToastHost } from './components/ToastHost';
 import { LoadingPlaceholder } from './components/LoadingPlaceholder';
+import { CourtCard } from './components/CourtCard';
+
+// Extracted utilities
+import { classForStatus, namesFor, formatTime, computeClock } from './utils/courtUtils.js';
 
 // Access shared utils from window for backward compatibility
 // (U now unused but kept for backward compatibility comment)
@@ -36,33 +40,6 @@ const readOnlyWrite = (...args) => {
   return false;
 };
 
-// Status → Tailwind classes (no color classes outside this map)
-function classForStatus(statusObj) {
-  const s = typeof statusObj === 'string' ? statusObj : statusObj?.status;
-  const selectable = statusObj?.selectable;
-
-  switch (s) {
-    case 'occupied':
-      return 'bg-gradient-to-b from-blue-300 to-blue-400 border-blue-300';
-    case 'overtime':
-      return selectable
-        ? 'bg-gradient-to-b from-emerald-600 to-emerald-700 border-emerald-600'
-        : 'bg-gradient-to-b from-blue-500 to-blue-600 border-blue-500';
-    case 'overtime-available':
-      return 'bg-gradient-to-b from-emerald-600 to-emerald-700 border-emerald-600';
-    case 'overtimeAvailable':
-      return 'bg-gradient-to-b from-green-600 to-green-700 border-green-600';
-    case 'wet':
-      return 'bg-gradient-to-b from-gray-300 to-gray-400 border-gray-300';
-    case 'blocked':
-      return 'bg-gradient-to-b from-slate-300 to-slate-400 border-slate-400';
-    case 'free':
-    default:
-      return 'bg-gradient-to-b from-emerald-400 to-emerald-500 border-emerald-400';
-  }
-}
-
-// Unified status source for the board UI
 // ---- Debounce helper (no UI change) ----
 const debounce = (fn, ms = 150) => {
   let t;
@@ -71,106 +48,6 @@ const debounce = (fn, ms = 150) => {
     t = setTimeout(() => fn(...a), ms);
   };
 };
-
-// ---- Helpers for domain-based tile text ----
-function namesFor(courtObj) {
-  if (Array.isArray(courtObj?.session?.group?.players)) {
-    return courtObj.session.group.players
-      .map((p) => p?.name)
-      .filter(Boolean)
-      .join(', ');
-  }
-  const last =
-    Array.isArray(courtObj?.history) && courtObj.history.length
-      ? courtObj.history[courtObj.history.length - 1]
-      : null;
-  if (Array.isArray(last?.players)) {
-    return last.players
-      .map((p) => p?.name)
-      .filter(Boolean)
-      .join(', ');
-  }
-  return '';
-}
-
-function formatTime(dt) {
-  if (!dt) return null;
-  const d = dt instanceof Date ? dt : new Date(dt);
-  if (isNaN(+d)) return null;
-  return d.toLocaleTimeString([], { hour: 'numeric', minute: '2-digit' });
-}
-
-function computeClock(status, courtObj, now, checkStatusMinutes = 150) {
-  // Helper to compute "+X min over" label
-  const getOvertimeLabel = (endTime) => {
-    if (!endTime) return 'Overtime';
-    const minutesOver = Math.round((now - endTime) / 60000);
-    return minutesOver > 0 ? `+${minutesOver} min over` : 'Overtime';
-  };
-
-  if (status === 'occupied') {
-    const end = courtObj?.session?.scheduledEndAt
-      ? new Date(courtObj.session.scheduledEndAt)
-      : null;
-    if (end && end > now) {
-      const mins = Math.max(0, Math.ceil((end - now) / 60000));
-      return {
-        primary: `${mins} min`,
-        secondary: `Until ${end.toLocaleTimeString([], { hour: 'numeric', minute: '2-digit' })}`,
-      };
-    }
-    return { primary: getOvertimeLabel(end), secondary: '' };
-  }
-  if (status === 'overtime') {
-    const end = courtObj?.session?.scheduledEndAt
-      ? new Date(courtObj.session.scheduledEndAt)
-      : null;
-    const start = courtObj?.session?.startedAt ? new Date(courtObj.session.startedAt) : null;
-    if (start) {
-      const minutesPlaying = Math.floor((now - start) / 60000);
-      if (checkStatusMinutes > 0 && minutesPlaying >= checkStatusMinutes) {
-        return {
-          primary: getOvertimeLabel(end),
-          secondary: 'check status',
-          secondaryColor: 'yellow',
-        };
-      }
-    }
-    return { primary: getOvertimeLabel(end), secondary: '' };
-  }
-  if (status === 'free') return { primary: 'Available', secondary: '' };
-  if (status === 'wet') return { primary: '🌧️\nWET COURT', secondary: '' };
-  if (status === 'blocked') {
-    const s = courtObj || {};
-    const rawLabel = s.blockedLabel || s.reason || 'Blocked';
-
-    let label = rawLabel.toUpperCase();
-    if (label.includes('COURT WORK') || label.includes('MAINTENANCE')) {
-      label = 'COURT WORK';
-    } else if (label.includes('LESSON')) {
-      label = 'LESSON';
-    } else if (label.includes('CLINIC')) {
-      label = 'CLINIC';
-    } else if (label.includes('LEAGUE')) {
-      label = 'LEAGUE';
-    }
-
-    const until = s.blockedEnd ? s.blockedEnd : null;
-    let secondary = '';
-    if (until) {
-      try {
-        const endTime = new Date(until);
-        if (!isNaN(endTime)) {
-          secondary = `Until ${endTime.toLocaleTimeString([], { hour: 'numeric', minute: '2-digit' })}`;
-        }
-      } catch (_e) {
-        // ignore invalid date
-      }
-    }
-    return { primary: label, secondary };
-  }
-  return { primary: '', secondary: '' };
-}
 
 // Timer registry for cleanup
 const _timers = [];
@@ -877,308 +754,6 @@ function TennisCourtDisplay() {
             courtBlocks={courtBlocks}
             upcomingBlocks={upcomingBlocks}
           />
-        </div>
-      </div>
-    </div>
-  );
-}
-
-// CourtCard Component
-function CourtCard({
-  courtNumber,
-  currentTime: _currentTime,
-  statusByCourt,
-  selectableByCourt,
-  statusObjectByCourt,
-  data,
-  isMobileView,
-  checkStatusMinutes = 150,
-  upcomingBlocks = [],
-  blockWarningMinutes = 60,
-  courts = [],
-  courtBlocks = [],
-}) {
-  const status = statusByCourt[courtNumber] || 'free';
-  const _selectable = selectableByCourt[courtNumber] || false; // Unused, kept for prop consistency
-  const statusObj = statusObjectByCourt?.[courtNumber] || {};
-  const cObj = data?.courts?.[courtNumber - 1] || {};
-
-  const now = new Date();
-  const {
-    primary,
-    secondary,
-    secondaryColor: _secondaryColor,
-  } = computeClock(status, status === 'blocked' ? statusObj : cObj, now, checkStatusMinutes);
-  const nm = namesFor(cObj);
-
-  const base =
-    'court-card border-4 rounded-xl flex flex-col items-center justify-start p-2 court-transition';
-  const courtClass = base + ' ' + classForStatus(statusObj);
-
-  // Mobile name formatting
-  function formatMobileNames(input) {
-    if (!input) return '';
-    const names = Array.isArray(input)
-      ? input
-      : String(input)
-          .split(/,\s*/)
-          .map((s) => s.trim())
-          .filter(Boolean);
-    if (!names.length) return '';
-
-    const SUFFIXES = new Set(['Jr.', 'Sr.', 'II', 'III', 'IV']);
-    const formatOne = (full) => {
-      const tokens = full.replace(/\s+/g, ' ').trim().split(' ');
-      if (tokens.length === 1) {
-        const t = tokens[0];
-        return t.length <= 3 ? t : `${t[0]}. ${t.slice(1)}`;
-      }
-      let last = tokens[tokens.length - 1];
-      let last2 = tokens[tokens.length - 2];
-      let lastName, remainder;
-      if (SUFFIXES.has(last)) {
-        lastName = `${last2} ${last}`;
-        remainder = tokens.slice(0, -2);
-      } else {
-        lastName = last;
-        remainder = tokens.slice(0, -1);
-      }
-      const first = remainder[0] || '';
-      const firstInitial = first ? `${first[0]}.` : '';
-      return `${firstInitial} ${lastName}`.trim();
-    };
-
-    const primaryName = formatOne(names[0]);
-    return names.length > 1 ? `${primaryName} +${names.length - 1}` : primaryName;
-  }
-
-  // Handler for occupied/overtime court taps (mobile only)
-  const handleOccupiedCourtTap = () => {
-    // Check overtime directly from session end time (not status, which uses different threshold)
-    const isOvertime =
-      cObj?.session?.scheduledEndAt && new Date(cObj.session.scheduledEndAt) < new Date();
-
-    if (!isMobileView) return;
-
-    // Check if court is overtime - treat like free court for registration
-    if (isOvertime) {
-      try {
-        // Check if empty playable courts exist - if so, block overtime tap
-        const playableCourts = listPlayableCourts(courts, courtBlocks, new Date().toISOString());
-        const emptyPlayable = playableCourts.filter((cn) => {
-          const c = courts[cn - 1];
-          return !c?.session;
-        });
-        if (emptyPlayable.length > 0) {
-          window.Tennis?.UI?.toast?.('Please select an available court', { type: 'warning' });
-          return;
-        }
-        // No empty courts - allow overtime takeover
-        window.mobileTapToRegister?.(courtNumber);
-        return;
-      } catch (e) {
-        console.error('[Overtime Tap] Error checking playable courts:', e);
-      }
-    }
-
-    // Court is truly occupied (not overtime) - handle clear court flow
-    const registeredCourt = sessionStorage.getItem('mobile-registered-court');
-
-    if (registeredCourt) {
-      // User has a registration - only allow clearing THEIR court
-      if (Number(registeredCourt) === courtNumber) {
-        window.MobileModal?.open('clear-court-confirm', { courtNumber });
-      }
-      // Tapping other occupied courts does nothing when registered
-    } else {
-      // No registration - show players on this court
-      window.MobileModal?.open('clear-court-confirm', {
-        courtNumber,
-        players: nm, // nm already has player names from namesFor(cObj)
-      });
-    }
-  };
-
-  const isOccupiedOrOvertime = status === 'occupied' || status === 'overtime';
-  const isClickable = status === 'free' || (isOccupiedOrOvertime && isMobileView);
-
-  return (
-    <div
-      className={courtClass}
-      data-court={courtNumber}
-      data-available={status === 'free'}
-      role={isClickable ? 'button' : undefined}
-      tabIndex={isClickable ? 0 : undefined}
-      onClick={
-        status === 'free'
-          ? () => window.mobileTapToRegister?.(courtNumber)
-          : isOccupiedOrOvertime && isMobileView
-            ? handleOccupiedCourtTap
-            : undefined
-      }
-      style={{ cursor: isClickable ? 'pointer' : 'default' }}
-    >
-      <h3
-        className={`court-text-lg font-bold ${status === 'wet' || status === 'blocked' ? 'text-gray-800' : 'text-white'} mb-1`}
-        style={{ textShadow: '1px 1px 2px rgba(0,0,0,0.3)' }}
-      >
-        Court {courtNumber}
-      </h3>
-      <div
-        className={`court-text-sm w-full ${status === 'wet' || status === 'blocked' ? 'text-gray-800' : 'text-white'} flex flex-col h-full`}
-      >
-        <div className="flex flex-col h-full w-full">
-          {status === 'free' ? (
-            <div className="relative flex-1 flex flex-col items-center justify-center w-full">
-              {/* Block warning - absolutely positioned at top */}
-              {(() => {
-                const blockWarning = getUpcomingBlockWarningFromBlocks(
-                  courtNumber,
-                  blockWarningMinutes,
-                  upcomingBlocks
-                );
-                if (!blockWarning || blockWarning.minutesUntilBlock >= blockWarningMinutes)
-                  return null;
-
-                return (
-                  <div
-                    className="absolute top-0 left-0 right-0 court-text-base font-bold leading-tight text-center"
-                    style={{ color: 'yellow' }}
-                  >
-                    {blockWarning.reason} in {blockWarning.minutesUntilBlock}m
-                  </div>
-                );
-              })()}
-              {/* "Available" - centered in full space */}
-              <div
-                className={`${isMobileView ? 'court-text-base font-medium leading-tight text-sm font-normal' : 'court-text-base font-medium leading-tight'} text-center`}
-              >
-                {isMobileView ? 'Tap to Select' : primary}
-              </div>
-            </div>
-          ) : status === 'wet' ? (
-            <div className="mt-1 court-text-base font-bold leading-tight text-center flex-1 flex items-center justify-center">
-              <div className="flex flex-col items-center">
-                <div style={{ fontSize: '2.5em', lineHeight: '1' }}>🌧️</div>
-                <div className="mt-1">WET COURT</div>
-              </div>
-            </div>
-          ) : status === 'blocked' ? (
-            <>
-              <div className="text-center flex-1 flex items-center justify-center mt-1 font-bold leading-tight blocked-court-label">
-                {(() => {
-                  let blockReason = statusObj.blockedLabel || statusObj.reason || 'Blocked';
-                  const upperReason = blockReason.toUpperCase();
-                  if (
-                    upperReason.includes('SHORT') ||
-                    upperReason.includes('COURT WORK') ||
-                    upperReason.includes('MAINTENANCE')
-                  ) {
-                    return 'Court work';
-                  }
-                  if (upperReason.includes('LESSON')) return 'Lesson';
-                  if (upperReason.includes('CLINIC')) return 'Clinic';
-                  if (upperReason.includes('LEAGUE')) return 'League';
-                  return blockReason.charAt(0).toUpperCase() + blockReason.slice(1).toLowerCase();
-                })()}
-              </div>
-              {statusObj.blockedEnd && !isMobileView && (
-                <div className="mt-auto text-sm text-gray-700 opacity-90 text-center">
-                  Until {formatTime(statusObj.blockedEnd)}
-                </div>
-              )}
-            </>
-          ) : status === 'occupied' || status === 'overtime' ? (
-            isMobileView ? (
-              <>
-                {nm ? (
-                  <div
-                    className="mt-1 court-text-sm font-medium text-center flex-1 flex items-center justify-center"
-                    style={{ textShadow: '1px 1px 2px rgba(0,0,0,0.3)' }}
-                  >
-                    <div className="truncate text-sm leading-tight px-1" title={nm}>
-                      {formatMobileNames(nm)}
-                    </div>
-                  </div>
-                ) : (
-                  <div className="flex-1"></div>
-                )}
-                <div className="mt-auto text-sm opacity-90 text-center">{primary}</div>
-              </>
-            ) : (
-              <>
-                {/* For occupied: show block warning instead of "X min" if applicable */}
-                {status === 'occupied' &&
-                  (() => {
-                    const blockWarning = getUpcomingBlockWarningFromBlocks(
-                      courtNumber,
-                      0,
-                      upcomingBlocks
-                    );
-                    if (blockWarning && blockWarning.minutesUntilBlock < blockWarningMinutes) {
-                      return (
-                        <div
-                          className="mt-1 court-text-base font-bold leading-tight text-center"
-                          style={{ color: 'yellow' }}
-                        >
-                          {blockWarning.reason} in {blockWarning.minutesUntilBlock}m
-                        </div>
-                      );
-                    }
-                    return (
-                      <div className="mt-1 court-text-base font-bold leading-tight text-center">
-                        {primary}
-                      </div>
-                    );
-                  })()}
-                {/* For overtime: show primary then block warning below */}
-                {status === 'overtime' && (
-                  <>
-                    <div className="mt-1 court-text-base font-bold leading-tight text-center">
-                      {primary}
-                    </div>
-                    {(() => {
-                      const blockWarning = getUpcomingBlockWarningFromBlocks(
-                        courtNumber,
-                        0,
-                        upcomingBlocks
-                      );
-                      if (!blockWarning || blockWarning.minutesUntilBlock >= blockWarningMinutes)
-                        return null;
-                      return (
-                        <div
-                          className="mt-1 court-text-base font-bold leading-tight text-center"
-                          style={{ color: 'yellow' }}
-                        >
-                          {blockWarning.reason} in {blockWarning.minutesUntilBlock}m
-                        </div>
-                      );
-                    })()}
-                  </>
-                )}
-                {nm ? (
-                  <div
-                    className="mt-1 court-text-sm font-medium text-center flex-1 flex items-center justify-center"
-                    style={{ textShadow: '1px 1px 2px rgba(0,0,0,0.3)' }}
-                  >
-                    {nm}
-                  </div>
-                ) : (
-                  <div className="flex-1"></div>
-                )}
-                {status === 'occupied' && cObj?.session?.scheduledEndAt && (
-                  <div className="mt-auto text-sm opacity-90 text-center">
-                    Until {formatTime(cObj.session.scheduledEndAt)}
-                  </div>
-                )}
-                {status === 'overtime' && secondary && (
-                  <div className="mt-auto text-sm text-center" style={{ color: 'yellow' }}>
-                    {secondary}
-                  </div>
-                )}
-              </>
-            )
-          ) : null}
         </div>
       </div>
     </div>
